To me and Reijo it seems the minus must not be there. Removed.
The state vector is

x ≡ [σ R X1 X2]

so we have

(dx/dε)[1:6,1:6] = dσ/dε

We can get this from

dx/dε = dx/dr dr/dε = inv(dr/dx) dr/dε ≡ (dr/dx) \ (dr/dε)

where we have used the chain rule and the (multidimensional) inverse function
theorem.

The minus sign in Joona's original calculation comes from defining the residual
as -r, instead of +r, where r is the residual function named g! in the code.
This sign difference will of course be compensated by a sign difference in the
corresponding dr/dx term (which is automatically computed by ForwardDiff,
so its sign is not immediately visible in the source code text), so the value
of the material Jacobian dσ/dε - which is the product of the two terms - will
be the same in either case.

The point is, the sign convention of the manually calculated drde term must
match the one used in the definition of g!. A sign mistake here is not caught
by testing the behavior of a single material point, because the material
Jacobian is only used by FEM calculations. (From the perspective of Materials.jl
itself, it's write-only.)