Add ChabocheOptions, so far with a single property: nlsolve_method
As per discussion with Reijo, it would be useful to be able to choose the method
used by NLsolve for solving the nonlinear equation system. Since the nonlinear
equation system is solved in a tight loop (especially considering FEM), this
allows the user to increase performance by choosing a faster method when that is
appropriate for the problem being solved.

TODO later: we should also think about which method to use as the default.
trust_region is NLsolve's own default. It is robust, but typically requires a
large number of evaluations of the residual function. Erring on the side of
caution is good numerics, so the practical question is: is there another method
that is at least equally robust, but provides higher performance?