Fix Julia 1.0 compatibility
In Julia 1.0, `I` is UniformScaling{Bool}, whereas (at least) Julia 1.3 and
later have added `I(n)` as shorthand for `Diagonal{Bool,Array{Bool,1}}`.

If you encounter the error:
    LoadError: MethodError: objects of type UniformScaling{Bool} are not callable
this is a one possible reason for that.

Worked around by defining a custom `eye` function that creates a `Diagonal`
populated with ones.